requirements.txt
streamlit
gtts
google-adk
google-generativeai
pillow
python-dotenv
pyttsx3
